+++ config/terraform/aws/variables.tf
@@ -104,6 +104,14 @@ variable "ecs_secret_backup_code_admin_email_template_id_fr" {
   type = string
 }
 
+variable "ecs_secret_confirmation_email_template_id_en" {
+  type = string
+}
+
+variable "ecs_secret_confirmation_email_template_id_fr" {
+  type = string
+}
+
 variable "ecs_secret_django_secret_key" {
   type = string
 }